requirements.txt
Python==3.8.10
requests==2.32.3
python-dotenv==1.0.1
